requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic-settings==2.7.1
apify-client>=1.6.0,<1.8
httpx==0.28.1
PyJWT==2.9.0
